server/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "dynamic_pricing"
version = "0.1.0"
description = "Pricing model endpoint server"

dependencies = [
    "psycopg2-binary==2.9.9",
    "pydantic==2.6.4",
    "pandas",
    "python-dotenv",
    "pydantic",
    "google-auth",
    "requests",
    "flask",
    "Jinja2",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
    "black",
    "pytest"
]
